train pessimistic runs
